# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  comment.Attachment:
    properties:
      filename:
        type: string
      id:
        type: string
      mimeType:
        type: string
      size:
        type: integer
      type:
        type: string
      uploadedAt:
        type: string
      url:
        type: string
    type: object
  comment.Comment:
    properties:
      attachments:
        items:
          $ref: '#/definitions/comment.Attachment'
        type: array
      authorAvatar:
        type: string
      authorEmail:
        type: string
      authorId:
        description: Author info
        type: string
      authorName:
        type: string
      content:
        description: Content
        type: string
      contentHtml:
        type: string
      createdAt:
        description: Timestamps
        type: string
      deletedAt:
        type: string
      depth:
        description: Depth for nested replies
        type: integer
      dislikeCount:
        type: integer
      id:
        type: string
      isAnonymous:
        type: boolean
      isDeleted:
        type: boolean
      isEdited:
        type: boolean
      isPinned:
        description: Features
        type: boolean
      likeCount:
        type: integer
      metadata:
        additionalProperties: {}
        description: Metadata
        type: object
      moderatedAt:
        type: string
      moderatedBy:
        type: string
      parentId:
        type: string
      pinnedAt:
        type: string
      reactionCounts:
        additionalProperties:
          type: integer
        type: object
      rejectionReason:
        type: string
      replyCount:
        description: Stats
        type: integer
      reportCount:
        type: integer
      resourceId:
        type: string
      resourceType:
        type: string
      rootId:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/comment.CommentStatus'
        description: Moderation
      tenantId:
        type: string
      updatedAt:
        type: string
    type: object
  comment.CommentStats:
    properties:
      approvedCount:
        type: integer
      pendingCount:
        type: integer
      reactionBreakdown:
        additionalProperties:
          format: int64
          type: integer
        type: object
      rejectedCount:
        type: integer
      totalComments:
        type: integer
    type: object
  comment.CommentStatus:
    enum:
    - pending
    - approved
    - rejected
    - spam
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusApproved
    - StatusRejected
    - StatusSpam
  comment.ListCommentsResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/comment.Comment'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  comment.RepliesResponse:
    properties:
      replies:
        items:
          $ref: '#/definitions/comment.Comment'
        type: array
      total:
        type: integer
    type: object
...
  github_com_minisource_feedback_internal_models.CreateCommentRequest:
    properties:
      content:
        maxLength: 2000
        type: string
      isAnonymous:
        type: boolean
      parentId:
        type: string
    required:
    - content
    type: object
...
  github_com_minisource_feedback_internal_models.UpdateCommentRequest:
    properties:
      content:
        maxLength: 2000
        type: string
    required:
    - content
    type: object
...
host: localhost:5012
info:
  contact: {}
  description: Feedback management service for Minisource
  title: Feedback Service API
  version: "1.0"
paths:
...
  /comments/{id}:
    delete:
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete comment
      tags:
      - Comments
    get:
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.Comment'
      summary: Get comment by ID
      tags:
      - Comments
    put:
      consumes:
      - application/json
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Update comment request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_minisource_feedback_internal_models.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.Comment'
      summary: Update comment
      tags:
      - Comments
  /comments/{id}/reactions:
    delete:
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Remove reaction from comment
      tags:
      - Comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Reaction request
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add reaction to comment
      tags:
      - Comments
  /comments/{id}/replies:
    get:
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.RepliesResponse'
      summary: Get replies to a comment
      tags:
      - Comments
...
  /feedback/{feedback_id}/comments:
    get:
      parameters:
      - description: Feedback ID
        in: path
        name: feedback_id
        required: true
        type: string
      - description: Parent comment ID
        in: query
        name: parent_id
        type: string
      - description: Sort by (created_at, like_count)
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: sort_order
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.ListCommentsResponse'
      summary: List comments
      tags:
      - Comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Feedback ID
        in: path
        name: feedback_id
        required: true
        type: string
      - description: Create comment request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_minisource_feedback_internal_models.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/comment.Comment'
      summary: Create comment
      tags:
      - Comments
  /feedback/{feedback_id}/comments/stats:
    get:
      parameters:
      - description: Feedback ID
        in: path
        name: feedback_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.CommentStats'
      summary: Get comment statistics
      tags:
      - Comments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"